* 'master' of https://github.com/CartoDB/mobile-sdk: (23 commits)
  Safety checks when drawing too large popups/texts
  C++11-fy, nullptr safety check
  xbuild to msbuild
  Get official Angle package from nuget for UWP builds
  Fix automatic UTF grid configuration when using raster tiles vis CartoVisLoader
  Minor API change: TextStyle.fontSize is now float (previously int)
  Introduced ‘RenderScale’ style attribute to labels. This can be used to increase sharpness/resolution of Text vector elements
  Fixed online style updater ignoring style.json changes
  Internal BinaryData API update
  Doc fixes and code formatting
  Improved VT click detection extra radius handling
  Optimization
  Optimization
  Added preliminary implementation for TileMask.getBoundingPolygon
  Minor tweak
  Fix batching (performance) issue with non-overlapping points
  Refactoring, file support for URLFileLoader
  Support for importing package manager packages directly from assets/network. Refactoring.
  Inline doc fix
  api.nutiteq.com -> mobile-api.carto.com
  ...

+++ all/native/styles/TextStyleBuilder.h
@@ -52,12 +52,12 @@ namespace carto {
          * Returns the font size for the text label.
          * @return The font size for the text label in points.
          */
-        int getFontSize() const;
+        float getFontSize() const;
         /**
          * Sets the font size for the text label.
          * @param size The new font size for the text label in points.
          */
-        void setFontSize(int size);
+        void setFontSize(float size);
 
         /**
          * Returns the stroke color for the text label.
@@ -92,7 +92,7 @@ namespace carto {
         
         std::string _textField;
 
-        int _fontSize;
+        float _fontSize;
         
         Color _strokeColor;
